add some feedback for fast mode
Sometimes the admin isn't there to watch the clock and a middle man
would tell how much the clock is off, but just once. Print how much the
clock has rolled while looping so that it can be interrupted at the
right moment. (A command line argument would be just too fancy.)